Move createSession middleware into SessionController
